# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gochive apis",
        "title": "gochive",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/amingze/gochive"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://github.com/amingze/gochive/blob/master/LICENSE"
        },
        "version": "0.0.1"
    },
    "basePath": "/api/",
    "paths": {
        "/matters": {
            "post": {
                "security": [
                    {
                        "OAuth2Application": [
                            "matter",
                            "admin"
                        ]
                    }
                ],
                "description": "根据文件信息上传,用于快速上传",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matters"
                ],
                "summary": "快速上传",
                "parameters": [
                    {
                        "description": "参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bind.BodyMatter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    }
                }
            }
        },
        "/matters/upload": {
            "post": {
                "security": [
                    {
                        "OAuth2Application": [
                            "matter",
                            "admin"
                        ]
                    }
                ],
                "description": "创建文件,用于快速上传",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matters"
                ],
                "summary": "创建文件",
                "parameters": [
                    {
                        "type": "string",
                        "name": "dir",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "fid",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "is_dir",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "signature",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    }
                }
            }
        },
...
        "/user/authorize": {
            "post": {
                "description": "创建秘钥",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAuthorize"
                ],
                "summary": "创建秘钥",
                "parameters": [
                    {
                        "description": "参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bind.BodyUserKeyCreation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    }
                }
            }
        },
        "/user/authorize/{name}": {
            "get": {
                "description": "查询秘钥",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAuthorize"
                ],
                "summary": "查询秘钥",
                "parameters": [
                    {
                        "type": "string",
                        "description": "秘钥名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除秘钥",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAuthorize"
                ],
                "summary": "删除秘钥",
                "parameters": [
                    {
                        "type": "string",
                        "description": "秘钥名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    }
                }
            }
        },
        "/user/authorize/{name}/secret": {
            "patch": {
                "description": "重置秘钥",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAuthorize"
                ],
                "summary": "重置秘钥",
                "parameters": [
                    {
                        "type": "string",
                        "description": "秘钥名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "bind.BodyMatter": {
            "type": "object",
            "properties": {
                "dir": {
                    "type": "string"
                },
                "fid": {
                    "type": "integer"
                },
                "is_dir": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "bind.BodyUserKeyCreation": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "httputil.JSONResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "/api/tokens",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "matter": " Grants matter access and write"
            }
        }
    }
}